scripts: west_commands: runners: fix west rtt shell interaction
When the west rtt command is executed, two things happen:
- A GDB server is run.
- A telnet client (via sockets) is opened to connect to that server.

Previously, that telnet client was operating in the line mode in both
directions. That mode is perfectly fine for just presenting the captured
content (RTT data) to the stdout; however, it is not suitable for
interacting with the device via the shell subsystem. None of the
control commands, such as tab and arrow keys, were correctly passed down
to the device.

To solve that the telnet client had to be switched into the character
mode. For that two separate implementations were needed, one for the
Unix and one for Windows. That was solved with two different context
managers.

They both wrap the common logic: two threads, where one reads input
from the keyboard and sends it to the telnet client, and the other on
reads the data from the telnet client and prints it to the screen.

Signed-off-by: Marko Sagadin <marko.sagadin42@gmail.com>